Extract shared CommandHelpers and PlatformRestrictable modules
Step and Migration duplicated five identical private helpers (command,
env_command, shell_script, command_succeeds?, command_exists?) and four
identical class-level platform DSL methods (macos_only, macos_only?,
debian_only, debian_only?). Extract them into Dotfiles::CommandHelpers
(included) and Dotfiles::PlatformRestrictable (extended), and include/
extend both in Step and Migration.

The included helpers are re-declared private in each base so they stay
out of the public interface. execute and allowed_on_platform? stay in
each base because they differ (Step#execute debugs; the two
allowed_on_platform? implementations differ).
